S302T300-81 #time 10m listing price now a Double ?fix

+++ backend/src/main/java/com/navbara_pigeons/wasteless/entity/Listing.java
@@ -1,7 +1,6 @@
 package com.navbara_pigeons.wasteless.entity;
 
 import com.navbara_pigeons.wasteless.dto.CreateListingDto;
-import java.time.LocalDate;
 import java.time.ZonedDateTime;
 import javax.persistence.CascadeType;
 import javax.persistence.Column;
@@ -12,7 +11,6 @@ import javax.persistence.GenerationType;
 import javax.persistence.Id;
 import javax.persistence.JoinColumn;
 import javax.persistence.ManyToOne;
-import javax.persistence.OneToOne;
 import javax.persistence.Table;
 import lombok.Data;
 
@@ -42,7 +40,7 @@ public class Listing {
   private long quantity;
 
   @Column(name = "PRICE")
-  private float price;
+  private Double price;
 
   @Column(name = "MORE_INFO")
   private String moreInfo;
